requirements.txt
asgiref==3.2.7
Django==2.2.5
django-filter==2.2.0
django-test-pep8==0.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
inflection==0.4.0
pep8==1.7.1
pi==0.1.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pytz==2020.1
sqlparse==0.3.1
